requirements.txt
streamlit
pillow
python-dotenv
openai>=1.0.0